android: Move SbAudioSink Init to CobaltActivity
Currently, `audio_sink::SbAudioSinkImpl::Initialize()` is called
from application_android.cc, which makes it intialize before the cobalt
browser client is able to initialize the feature list and field trials
parameters (this initialization happens in cobalt_main_delegate.cc).
This order of initialization makes it impossible to set up any starboard
experiments for AudioTrackAudioSink, as the app will crash due to the
AudioTrackAudioSink requesting information from the starboard
FeatureList before it is initialized.

This change moves the `audio_sink::SbAudioSinkImpl::Initialize()`
function to `CobaltActivity.java` in the `finishInitialization`
function. This change is performed by adding a JNI function
`InitializePlatformAudioSink` to `StarboardBridge.java` which is called
in CobaltActivity. This function is then linked to `starboard_bridge.cc`
which calls the initialize method.

Bug: 438001753